version check IMGDAT; 64bit tweaks; new sprites/palettes

+++ palmod/Game/MVC2_PALSEL.cpp
@@ -1336,7 +1336,7 @@ BOOL CGame_MVC2_D::UpdatePalImg(int Node01, int Node02, int Node03, int Node04)
         UINT16 nSrcAmt = 1;
         UINT16 nNodeIncrement = 1;
 
-        UINT16 nTeamViewNode = (UINT16)floor(NodeGet->uPalId / ARRAYSIZE(DEF_BUTTONLABEL6_MVC2));
+        UINT16 nTeamViewNode = (UINT16)floor(NodeGet->uPalId / (UINT16)ARRAYSIZE(DEF_BUTTONLABEL6_MVC2));
         const sDescTreeNode* pCurrentNode = &MVC2_A_TEAMVIEW_COLLECTION[nTeamViewNode];
 
         for (UINT16 nTeamIndex = 0; nTeamIndex < ARRAYSIZE(mvc2TeamList); nTeamIndex++)
